paging bug fix and keyboard interrupt handler and scancode translator

+++ kernel/src/math_util.h
@@ -2,6 +2,6 @@
 
 struct Point
 {
-    unsigned int x;
-    unsigned int y;
+    long x;
+    long y;
 };
